Update to use git switch instead of git checkout
The new commands `git switch`and `git restore` were introduced
in 2.23 back in August 2019, and are much cleaner than the
old `git checkout`, so I believe that we
should be teaching the new way, not the old.

While `git switch` and `git restore` are in principal still marked as
experimental in the documentation, they are here to stay.
Even output from commands like git status,
have gone away from suggesting `git checkout`.

This commit goes through all the README files and makes appropriate
changes to recomend `git switch` instead of `git checkout`
(and `git switch -c` instead of `git checkout -b`)

The actual setup scripts are still left untouched for maximum compatablility.
If you have e.g. participants in a training using very old Git versions,
it is still a lot easier to tell them to use `git checkout`
when `git switch`is mentioned,
than it is to debug scripts that are failing miserably.