Fixed typo in defaultBranch config #319
When creating new bare repo for fake remote usage, default branch name is
supposed to be set according to variable declared in utils.sh
The git command that creates the repo has the command misspelled and hence
the branch name isn't set correctly. This is only noticeable if the user has
changed their default branch name config to something else than master.

This patch fixes the typo and the config should now work.